backend/requirements.txt
fastapi
uvicorn
beanie
motor
python-jose[cryptography]
passlib[bcrypt]
pydantic